refactor: remove local workspace mode from agents/chat
